+++ apps/frontend/src/pages/media/commit.tsx
@@ -4,6 +4,7 @@ import { useCommitMedia } from "@/lib/hooks/graphql";
 import LoadingPage from "@/lib/layouts/LoadingPage";
 import LoggedIn from "@/lib/layouts/LoggedIn";
 import { getLot } from "@/lib/utilities";
+import type { MetadataSource } from "@ryot/generated/graphql/backend/graphql";
 import Head from "next/head";
 import { useRouter } from "next/router";
 import { type ReactElement, useEffect } from "react";
@@ -12,13 +13,15 @@ const Page: NextPageWithLayout = () => {
 	const router = useRouter();
 	const lot = getLot(router.query.lot);
 	const identifier = router.query.identifier?.toString();
+	const source = router.query.source?.toString() as unknown as MetadataSource;
 
 	const commitMedia = useCommitMedia(lot, (id) => {
 		router.replace(`${ROUTES.media.details}?item=${id}`);
 	});
 
 	useEffect(() => {
-		if (identifier && lot) commitMedia.mutate({ identifier, lot });
+		if (identifier && lot && source)
+			commitMedia.mutate({ identifier, lot, source });
 	}, [identifier, lot]);
 
 	return (